Bug #12405: fix "MetaDataException to be thrown, but nothing was thrown"
The code was comparing the unitsScrollLocalDate (that is a LocalDate in UTC) to LocalDate.now() (that is the current local date **in local timezone**).
LocalDate.now() could be UTC day + 1 at the end of the day (after 11pm for UTC+1 TZ).
By using LocalDateUtil.now().toLocalDate(), we get the LocalDate in UTC (as LocalDateUtil.now() use a clock in UTC), which should fix the issue.